% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSim.R
\name{calcSeeds}
\alias{calcSeeds}
\title{Calculate Seeds}
\usage{
calcSeeds(ms, seeds, filename = NULL, allowDeletion = FALSE)
}
\arguments{
\item{ms}{object of class \code{multSim} (with existing simulations) that 
shall be expanded}

\item{seeds}{integer vector with seeds that shall be simulated
(different seeds lead to different simulated trajectories)}

\item{filename}{string which indicates the path where to save the result, 
i. e. "NameOfFile.rda". In case of a break or an error, already simulated 
seeds will still be saved there. If \code{filename} = NULL, the result will
not be saved in a file.}

\item{allowDeletion}{logical. If true, seeds with existing simulations 
stored in \code{ms} will be deleted in case they don't appear in 
the new vector of seeds.}
}
\value{
object of class "multSim" containing all simulations
}
\description{
Simulates all seeds in a given vector that are not simulated yet.
}
\seealso{
multSim
}
\keyword{internal}
