% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSimData_outputMethods.R,
%   R/multSim_outputMethods.R
\name{density}
\alias{density}
\alias{density.multSimData}
\alias{density.multSim}
\title{Plot Density Estimations}
\usage{
\method{density}{multSimData}(x, t, main, sub, ...)

\method{density}{multSim}(x, t, main, ...)
}
\arguments{
\item{x}{object of class \code{\link{multSimData}} or \code{\link{multSim}}.}

\item{t}{a vector of time values at which the densities shall be plotted. 
If no vector is provided, every time value provided will
be plotted.}

\item{main}{optional character string for the title of the plot. If \code{x} 
is a \code{multSim} object, \code{main} will be set as \code{descr(x$model)} 
if not otherwise specified.}

\item{sub}{optional character string for the subtitle of the plot. The 
default value for a \code{multSim} object \code{x} gives informations about 
parameters and the initial values.}

\item{...}{additional parameters passed to the default method of 
\code{\link[stats]{density}}}
}
\description{
A plot method for simulations of a piecewise deterministic 
markov process (\code{\link{PDMP}}). It computes and plots a density over 
all simulations of the PDMP, seperately for every time value and every 
contious variable. Discrete variables are plotted in a stacked barplot.
}
\examples{
data("simplePdmp")
ms <- multSim(simplePdmp, seeds = 1:10)
density(ms, t = c(5, 10))
}
