% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSim.R
\name{expandTimes}
\alias{expandTimes}
\title{Expand simulation time}
\usage{
expandTimes(ms, to, filename = NULL)
}
\arguments{
\item{ms}{object of class "multSim" (with existing simulations) that shall 
be expanded (simulations are in ms$outputList; every NA in ms$outputList 
will be skipped and not simulated)}

\item{to}{new stopping time, should be larger than times(ms$model)["to"]}

\item{filename}{string which indicates the path where to save the result, 
i. e. "NameOfFile.rda". In case of a break or an error, already simulated 
seeds will still be saved there. If \code{filename} = NULL, the result will
not be saved in a file.}
}
\value{
object of class "multSim" containing the simulations
}
\description{
Expands all existing simulations to a longer time period
}
\seealso{
multSim
}
\keyword{internal}
