% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjp_methods.R
\docType{methods}
\name{format,mjpModel-method}
\alias{format,mjpModel-method}
\alias{format}
\title{Format a MJP for pretty printing.}
\usage{
\S4method{format}{mjpModel}(x, begin = NULL, end = NULL, short = TRUE,
  slots = c("parms", "init", "times"), sep, collapse)
}
\arguments{
\item{x}{an object of class \code{\link{mjpModel}} or one of its subclasses.}

\item{begin}{a character string that is pasted at the beginning.}

\item{end}{a character string that is pasted at the end.}

\item{short}{logical. If TRUE, a shorter version without space characters
will be returned (useful to generate filenames). Defaults to TRUE.}

\item{slots}{a vector specifying the names of all slots that shall be pasted
to the string. ' The default is \code{c("parms", "init", "times")}. 
Supported slots are "descr", "parms", "init",  and "times".}

\item{sep}{a character string to separate the names and values of slots that 
are vectors or lists (e. g. "times" or "parms"). It defaults to \code{"="} 
if \code{short} is \code{TRUE} and to \code{" = "} if \code{short} is 
\code{FALSE}.}

\item{collapse}{a character string to separate the slots.
It defaults to \code{"___"} if \code{short} is \code{TRUE} and to \code{". "} 
if \code{short} is \code{FALSE}.}
}
\description{
Give important information about a \code{\link{mjpModel}} in one character 
string. This method is used internally to generate default filenames or plot 
titles.
}
\examples{
data("KendallBD")
format(KendallBD, begin = "ToggleSwitch__", end = ".rda")
format(KendallBD, begin = paste0(descr(toggleSwitch), ": "), short=FALSE)
parms(KendallBD) <- list()
cat(format(KendallBD, short = FALSE, collapse = ".\\n",
           slots = c("init", "times","parms"),
           begin = "Kendalls birth-death-process:\\n"))

}
