% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSimData_outputMethods.R,
%   R/multSim_outputMethods.R
\name{hist}
\alias{hist}
\alias{hist.multSimData}
\alias{hist.multSim}
\title{Histogram over all simulations}
\usage{
\method{hist}{multSimData}(x, t, main, sub, ...)

\method{hist}{multSim}(x, t, main, ...)
}
\arguments{
\item{x}{object of class \code{\link{multSimData}} or \code{\link{multSim}}.}

\item{t}{a single time value at which the histogram shall be plotted.
The parameter can be omitted if \code{x} contains only one time value.}

\item{main}{optional character string for the title of the plot. If \code{x} 
is a \code{multSim} object, \code{main} will be set as \code{descr(x$model)} 
if not otherwise specified.}

\item{sub}{optional character string for the subtitle of the plot. The 
default value for a \code{multSim} object \code{x} gives informations about 
parameters and the initial values.}

\item{...}{additional parameters passed to the default method of 
\code{\link[graphics]{hist}}}
}
\description{
A plot method for simulations of a piecewise deterministic markov process 
(\code{\link{PDMP}}). It plots every continous variable in its own histogram 
and all discrete variables in a stacked barplot.
}
\examples{
data("simplePdmp")
ms <- multSim(simplePdmp, seeds = 1:10)
hist(ms, t = 10)
hist(getMultSimData(ms), t = 10, density = 10)
}
