% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IASP}
\alias{IASP}
\title{An age-structured  process with aging and imigration}
\format{An object of class \code{\link{pdmp_vd_Model}}.}
\usage{
IASP
}
\description{
This is a simple example for a variable dimension  process
defined as \code{\link{pdmp_vd_Model}}. It is included to the package for 
demonstration purposes.
}
\section{Slots}{

\describe{
\item{\code{parms}}{imigration rate.}

\item{\code{init}}{There is one real variable with initial value 0.}

\item{\code{jumpfunc}}{imigrations   jump from \code{x} to 
\code{c(0,x)}, meaning a newborn is added to the population at a constant rate.}

\item{\code{ratefunc}}{A vector of length two determining the probability of a jump 
being a birth or a death.}

\item{\code{times}}{The simulations will start at time \code{t = 0} and end at 
\code{t = 10} with step length 0.01.}
}}

\examples{
## the code used to generate this model:
IASP <- pdmp_vd_Model(
  descr = "a PDMP model for imigration and aging",
  init = c(0),
  times = c(from = 0, to = 1e2, by = 1),
  dynfunc = function(t, x, parms) 0*x+1,
  ratefunc = function(t, x, parms) parms[1],
  jumpfunc = function(t, x, parms, jtype) c(0,x),
  summaryfunc= function(x) c(mean=mean(x),size=length(x)),
  parms = c(lambda=1)
)

## load it and plot a simulation:
data("IASP")
plot(sim(IASP))
}
\seealso{
\code{\link{pdmp_vd_Model-class}}
for the formal description of the S4 class.
}
\keyword{datasets}
