% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msCsv.R
\name{loadMultSimCsv}
\alias{loadMultSimCsv}
\alias{loadmultsimcsv}
\alias{loadMultSimCsv.multSimCsv}
\alias{loadMultSimCsv.character}
\title{load csv files with simulation results}
\usage{
loadMultSimCsv(x, dir)

\method{loadMultSimCsv}{multSimCsv}(x, dir = ".")

\method{loadMultSimCsv}{character}(x, dir = ".")
}
\arguments{
\item{x}{either an object  of class \code{\link{multSimCsv}} or a character 
string indicating the filename of a stored multSimCsv object, e. g. 
"exampleModel_MultSimCsv.rda"}

\item{dir}{directory where the csv files are stored. 
Defaults to the working directory.}
}
\value{
object of class \code{\link{multSimCsv}}
}
\description{
Load an object of class \code{multSimCsv} created with method
\code{\link{multSimCsv}} from a directory and create LaF links to
the stored csv files (\pkg{LaF} is a package that provides a fast access
to data stored in csv files). The method returnes an object of class
\code{multSimCsv} with the LaF links stored in element \code{lafList}.
}
\section{Methods (by class)}{
\itemize{
\item \code{multSimCsv}: create laf-links for multSimCsv objects

\item \code{character}: load multSimCsv from rda file and create LaF links
}}

\note{
Opening the saved rda file directly with \code{\link{readRDS}} will 
lead to a \code{multSimCsv} object with \code{lafList} set to NULL.
The links have to be created during run-time which is the reason why
method \code{loadMultSimCsv} exists.
}
\seealso{
\code{\link{multSimCsv}} to create the csv files, 
\code{vignette("LaF-manual")} for an introduction of package \pkg{LaF}
}
