% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msCsv_calculations.R, R/multSim_calculations.R
\name{mean.multSimCsv}
\alias{mean.multSimCsv}
\alias{mean}
\alias{mean.multSim}
\title{Methods for function \code{mean} in Package \pkg{pdmpsim}}
\usage{
\method{mean}{multSimCsv}(x, ...)

\method{mean}{multSim}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{multSim}} or \code{\link{multSimCsv}}}

\item{...}{additional arguments passed to \code{\link[base]{mean}}}
}
\value{
data.frame with calculated mean values
}
\description{
This method calculates the mean for every time value over all simulations.
}
