% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjp_accessors.R
\docType{methods}
\name{mjp-accessors}
\alias{mjp-accessors}
\alias{ratefunc,mjpModel-method}
\alias{jumpfunc,mjpModel-method}
\alias{descr,mjpModel-method}
\alias{ratefunc<-,mjpModel-method}
\alias{jumpfunc<-,mjpModel-method}
\alias{parms<-,mjpModel-method}
\alias{descr<-,mjpModel-method}
\alias{main<-,mjpModel-method}
\alias{equations<-,mjpModel-method}
\alias{observer<-,mjpModel-method}
\title{Accessor functions for Class mjpModel}
\usage{
\S4method{ratefunc}{mjpModel}(obj)

\S4method{jumpfunc}{mjpModel}(obj)

\S4method{descr}{mjpModel}(obj)

\S4method{ratefunc}{mjpModel}(obj) <- value

\S4method{jumpfunc}{mjpModel}(obj) <- value

\S4method{parms}{mjpModel}(obj) <- value

\S4method{descr}{mjpModel}(obj) <- value

\S4method{main}{mjpModel}(obj) <- value

\S4method{equations}{mjpModel}(obj) <- value

\S4method{observer}{mjpModel}(obj) <- value
}
\arguments{
\item{obj}{an object of class \code{\link{mjpModel}}}

\item{value}{the value that shall be set}
}
\description{
Get or set the slot values for an object of class \code{\link{mjpModel}}. 
In most cases, setting a new value also affects the slot \code{out} 
(where the simulation result can be stored). This applies to slots 
\code{times}, \code{parms}, \code{init}, \code{ratefunc},
\code{jumpfunc} and \code{initfunc}, which set out to NULL.
Setting a new value for \code{descr} will not affect 
the slot \code{out}.
}
\details{
It is also possible to modify the slots directly,
e.g. the parameters of a model \code{mjp} with \code{mjp@parms}, 
but this is normally not recommended as there is no guarantee that this 
will work in a compatible way in future versions.
}
\seealso{
Class definition \code{\link{mjpModel}}
}
