% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjp_methods.R
\docType{methods}
\name{mjpModel-methods}
\alias{mjpModel-methods}
\alias{format,mjpModel-method}
\alias{plot,mjpModel,missing-method}
\title{Format a MJP for pretty printing.}
\usage{
\S4method{format}{mjpModel}(x, begin = NULL, end = NULL,
  short = TRUE, slots = c("parms", "init", "times"), sep, collapse)
<<<<<<< HEAD:man/format-mjpModel-method.Rd
=======

\S4method{plot}{mjpModel,missing}(x, y, ...)
>>>>>>> e1dd096c41a1171ad9e8eed209379b09142abdeb:man/mjp-methods.Rd
}
\arguments{
\item{x}{an object of class \code{\link{mjpModel}} or one of its subclasses.}

\item{begin}{a character string that is pasted at the beginning.}

\item{end}{a character string that is pasted at the end.}

\item{short}{logical. If TRUE, a shorter version without space characters
will be returned (useful to generate filenames). Defaults to TRUE.}

\item{slots}{a vector specifying the names of all slots that shall be pasted
to the string. ' The default is \code{c("parms", "init", "times")}. 
Supported slots are "descr", "parms", "init",  and "times".}

\item{sep}{a character string to separate the names and values of slots that 
are vectors or lists (e. g. "times" or "parms"). It defaults to \code{"="} 
if \code{short} is \code{TRUE} and to \code{" = "} if \code{short} is 
\code{FALSE}.}

\item{collapse}{a character string to separate the slots.
It defaults to \code{"___"} if \code{short} is \code{TRUE} and to \code{". "} 
if \code{short} is \code{FALSE}.}

\item{y}{ignored}

\item{...}{optional plotting parameters}

\item{x}{an object of class mjpModel with a simulation 
stored in slot \code{out}}
}
\description{
Give important information about a \code{\link{mjpModel}} in one character 
string. This method is used internally to generate default filenames or plot 
titles.

This is method plots single simulations of
markov jump processes defined as
\code{\link{mjpModel}}. There are also other plot
methods available that use \pkg{ggplot2}.
}
\examples{
data("KendallBD")
format(KendallBD, begin = "ToggleSwitch__", end = ".rda")
format(KendallBD, begin = paste0(descr(toggleSwitch), ": "), short=FALSE)
parms(KendallBD) <- list()
cat(format(KendallBD, short = FALSE, collapse = ".\\n",
           slots = c("init", "times","parms"),
           begin = "Kendalls birth-death-process:\\n"))
data("SIRstoch")
sim <- sim(SIRstoch, seed = 1)
plot(sim, col = "red", lwd = 2)

# Alternative: plotSeeds
msim <- multSim(SIRstoch, seeds = 1)
plot <- plotSeeds(msim)
plot + ggplot2::facet_grid(variable ~ seed)
}
\seealso{
\code{\link{plotSeeds}} for another plot function
to plot single simulations
}
