% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjp_sim.R
\docType{methods}
\name{mjp.sim}
\alias{mjp.sim}
\alias{sim,mjpModel-method}
\alias{sim}
\alias{sim,mjpModel,ANY-method}
\alias{sim,mjpModel,mjpModel-method}
\title{Simulation of a mjpModel object}
\usage{
\S4method{sim}{mjpModel}(obj, initialize = FALSE, seed = 1, njump = 1e+06,
  outSlot = TRUE, ...)
}
\arguments{
\item{obj}{obj of class \code{\link{mjpModel}} or one of its subclasses}

\item{initialize}{boolean variable. If initialize = TRUE and \code{obj}
contains a user-defined initializing function (\code{initfunc}), this 
function will be called before the simulation. This can be useful i. e. 
to set a random initial value (see examples). If random numbers are 
generated within \code{initfunc}, they will be affected by the given 
\code{seed}. To avoid this, add \code{set.seed(NULL)} to the function
definition of \code{initfunc}.
Parameter \code{initialize} defaults to FALSE, because it is 
not necessary for simulating piecewise deterministic Markov models.}

\item{seed}{an integer or NULL. This makes the result of \code{sim} 
reproducible, although random numbers are used during execution. 
Simulation with equal seeds will lead to equal results. If seed is set to 
NULL, a new one is created from the current time and the process ID.}

\item{njump}{number of maximal possible jumps. The default value is 1e+06 
and should be sufficient. If the simulation breaks before the end of 
simulation time, try to set a higher value for \code{njump}.}

\item{outSlot}{boolean variable. If FALSE, only the result of the simulation 
is returned. If TRUE, the whole \code{obj} is returned with the simulation 
result stored in slot \code{out}.}

\item{...}{optional parameters passed to the solver function.}
}
\value{
The returned value depends on the parameter \code{outSlot}. 
If \code{outSlot = TRUE}, the function returns the complete pdmpModel 
instance with the simulation result saved in the \code{out} slot.
Otherwise, only the simulation result is returned.
}
\description{
Simulation of a mjpModel object
}
\note{
If the result is stored in slot \code{out}, it can get lost if the 
value of the other slots is changed via <-. 
See \code{\link{mjp-accessors}} for further informations.
}
\examples{
dontshow{
  require("simecol")
}

data("KendallBD") # load object of class pdmpModel

#### simulations

# simulate and plot the result
plot(sim(KendallBD, seed = 1))

# simulate and plot the result with random seed
plot(sim(KendallBD, seed = NULL))

# simulating again with seed = 1 leads to the same result as before
plot(sim(KendallBD, seed = 1))

#### be careful about slot out!

out(KendallBD) # NULL, because results have not been stored

simBD <- sim(kendallBD) # simulation results are stored in out
head(out(simBD))

init(KendallBD) <- c(N = 10)
head(out(KendallBD)) # NULL, because slot init has changed

simBD <- sim(KendallBD, outSlot = FALSE)
str(simBD) # only the simulation result, the pdmpModel object is lost


}
\seealso{
function \code{\link{multSim}} or \code{\link{multSimCsv}} 
for multiple simulations, ... for plot and summary methods of the simulation.
}
