% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSim_calculations.R
\name{moments}
\alias{moments}
\alias{moments.multSim}
\title{Calculate raw moments}
\usage{
moments(x, order)

\method{moments}{multSim}(x, order)
}
\arguments{
\item{x}{object of class \code{\link{multSim}} or \code{\link{multSimCsv}}}

\item{order}{number that specifies the order of the moments}
}
\value{
data.frame with calculated moments
}
\description{
This method calculates the raw moment of a given order for every time 
over all simulations.
}
