% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSim.R
\name{multSim}
\alias{multSim}
\title{Multiple simulations of a pdmp model}
\usage{
multSim(obj, seeds, filename = NULL, ms = NULL,
  allowDeletion = FALSE, ...)
}
\arguments{
\item{obj}{object of class \code{\link{pdmpModel}} or one of its subclasses}

\item{seeds}{integer vector with seeds that shall be simulated
(different seeds lead to different simulated trajectories)}

\item{filename}{string which indicates the path where to save the result, 
i. e. "NameOfFile.rda". In case of a break or an error, already simulated 
seeds will still be saved there. If \code{filename} = NULL, the result will
not be saved in a file.}

\item{ms}{object of class \code{multSim} (with existing simulations) that 
shall be expanded}

\item{allowDeletion}{logical. If true, seeds with existing simulations 
stored in \code{ms} will be deleted in case they don't appear in 
the new vector of seeds.}

\item{...}{additional parameters for method \code{\link{sim}}}
}
\value{
object of class \code{multSim} containing simulations for all 
given seeds
}
\description{
Perform simulations of a PDMP with different seeds and optionally save 
the results in a rda file.
}
\details{
The returned s3 class \code{multSim} contains 4 different elements:
\itemize{
\item \code{model} containing the pdmpModel \code{obj},
\item \code{seeds} containing the integer vector \code{seeds},
\item a list \code{outputList} containing the simulation results,
i.e. \code{outputList[[i]]} = \code{sim(model, seed = seeds[i])},
\item a list \code{timeList} containing the time needed for every simulation,
i.e. \code{timeList[[i]]} = \code{system.time(sim(model, seed = seeds[i]))}
}
If an object \code{ms} of class \code{multSim} is given as parameter, method 
\code{multSim} will only simulate the seeds that are not simulated yet. 
You can control if seeds that are already simulated and stored in \code{ms} 
but do not appear in the parameter \code{seeds} should be deleted or not. 
This is done via the parameter \code{allowDeletion}.

If the \code{time} slot of \code{obj} has a larger end value than the 
pdmpModel stored in \code{ms}, i.e. \code{times(obj)["to"] >} 
\code{times(ms$model)["to"]}, all existing simulations will be expanded to 
the larger end value.

Slot \code{obj@out} is not affected by \code{multSim}.
}
\examples{
data("simplePdmp")
m1 <- multSim(simplePdmp, seeds = 1:10)
plot(m1$outputList[[1]])

# expand times:
times(simplePdmp)["to"] <- 20
m2 <- multSim(simplePdmp, seeds = 1:15, ms = m1)
plot(m2$outputList[[15]])
}
\seealso{
There are several plot methods for objects of class \code{multSim}:
\code{\link[=plot.multSim]{plot}}, \code{\link{plotSeeds}}, \code{\link{plotTimes}},
\code{\link{plotStats}}, \code{\link{hist}} and \code{\link{density}}.
}
