% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msCsv.R
\name{multSim2multSimCsv}
\alias{multSim2multSimCsv}
\alias{multsim2multsimcsv}
\title{Convert 'multSim' to 'multSimCsv'}
\usage{
multSim2multSimCsv(ms, prefix = format(ms$model), append = FALSE,
  digits = NULL)
}
\arguments{
\item{ms}{object of class \code{\link{multSim}}}

\item{prefix}{string which determines the begin of all filenames that are 
created by the method, e.g. "NameOfModel". You can also specify a path where 
to save the result, e.g. "home/path/to/simulations/NameofModel".}

\item{append}{logical. If there already exists a file with the same filename 
and \code{append = TRUE}, the simulated results will be appended at the end 
of the file. Otherwise an error will be thrown.}

\item{digits}{integer indicating how many significant digits are to be saved 
in the csv files. Restricting digits can lead to smaller csv files. The 
default, NULL, uses \code{\link{getOption}("digits")}.}
}
\value{
object of class \code{\link{multSimCsv}} providing the filenames of 
all csv files that contain the simulated results with corresponding \pkg{LaF} 
objects stored in element \code{lafList}.
}
\description{
Convert an object of class \code{\link{multSim}} to an object of class 
\code{\link{multSimCsv}}. This can be useful if it turns out that the 
\code{\link{multSim}} object needs to much working memory.
}
