% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSimData.R
\docType{methods}
\name{multSimData}
\alias{multSimData}
\alias{getMultSimData}
\alias{getmultsimdata}
\alias{multsimdata}
\alias{getMultSimData.multSim}
\alias{getMultSimData.multSimCsv}
\alias{getMultSimData.multSimData}
\alias{getMultSimData,pdmpModel-method}
\title{Create data.frame with simulations}
\usage{
getMultSimData(x, times, seeds)

getMultSimData.multSim(x, times, seeds)

getMultSimData.multSimCsv(x, times, seeds)

getMultSimData.multSimData(x, times, seeds)

\S4method{getMultSimData}{pdmpModel}(x, times, seeds)
}
\arguments{
\item{x}{an object of class \code{\link{multSim}}, \code{\link{multSimCsv}},
\code{multSimData} or \code{\link{pdmpModel}}}

\item{times}{vector with specific time values for which the simulation
results (for all seeds) shall appear in the data.frame}

\item{seeds}{vector with specific seeds for which the simulation
results (for all times) shall appear in the data.frame}
}
\value{
a data.frame of class \code{multSimData} with simulation results.
}
\description{
This method transforms objects of class \code{multSim} or \code{multSimCsv}
into a specific data.frame of class \code{multSimData}. Most plot method
of package \pkg{pdmpsim} are based on this class. The data.frame has five
columns:
\itemize{
\item \code{time} for the time value,
\item \code{seed} for the seed on which the simulation was created,
\item \code{type} indicating if the variable is discrete or continous,
\item \code{variable} for the name of the simulated variable,
\item \code{value} for the simulated value.} 
Methods that use objects of class \code{multSimData} as input are
\code{\link[=plot.multSimData]{plot}}, \code{\link{plotSeeds}}, \code{\link{plotTimes}},
\code{\link{plotStats}}, \code{\link{hist}} and \code{\link{density}}.
}
\examples{
data("toggleSwitch")
ms <- multSim(toggleSwitch, seeds = 1:5)
d <- getMultSimData(ms, seeds = 1:3, times = c(5, 10))
print(d)
}
