% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmp_vd_accessors.R
\docType{methods}
\name{pdmp_vd-accessors}
\alias{pdmp_vd-accessors}
\alias{summaryfunc}
\alias{summaryfunc<-}
\alias{summaryfunc,pdmp_vd_Model-method}
\alias{summaryfunc<-,pdmp_vd_Model-method}
\title{Accessor functions for Class pdmp_vd_Model}
\usage{
summaryfunc(obj)

summaryfunc(obj) <- value

\S4method{summaryfunc}{pdmp_vd_Model}(obj)

\S4method{summaryfunc}{pdmp_vd_Model}(obj) <- value
}
\arguments{
\item{obj}{an object of class \code{\link{pdmp_vd_Model}}}

\item{value}{the value that shall be set}
}
\description{
Get or set the slot values for an object of class \code{\link{pdmp_vd_Model}}. 
In most cases, setting a new value also affects the slot \code{out} 
(where the simulation result can be stored). This applies to slots 
\code{times}, \code{parms}, \code{init}, \code{dynfunc}, \code{ratefunc},
\code{jumpfunc} and \code{initfunc}, which set out to NULL.
Setting a new value for \code{descr} or \code{discStates} will not affect 
the slot \code{out}.
}
\details{
It is also possible to modify the slots directly,
e.g. the parameters of a model \code{pdmp} with \code{pdmp@parms}, 
but this is normally not recommended as there is no guarantee that this 
will work in a compatible way in future versions.
}
\seealso{
Class definition \code{\link{pdmp_vd_Model}}
}
