% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmpBorder_accessors.R
\docType{methods}
\name{pdmpBorder-accessors}
\alias{pdmpBorder-accessors}
\alias{borderfunc}
\alias{borderfunc<-}
\alias{borroot}
\alias{terroot}
\alias{borderfunc,pdmpBorder-method}
\alias{terroot,pdmpBorder-method}
\alias{borroot,pdmpBorder-method}
\alias{borderfunc<-,pdmpBorder-method}
\alias{borroot<-,pdmpBorder-method}
\alias{terroot<-,pdmpBorder-method}
\title{Accessor functions for Class pdmpBorder}
\usage{
borderfunc(obj)

borderfunc(obj) <- value

borroot(obj)

terroot(obj)

\S4method{borderfunc}{pdmpBorder}(obj)

\S4method{terroot}{pdmpBorder}(obj)

\S4method{borroot}{pdmpBorder}(obj)

\S4method{borderfunc}{pdmpBorder}(obj) <- value

\S4method{borroot}{pdmpBorder}(obj) <- value

\S4method{terroot}{pdmpBorder}(obj) <- value
}
\arguments{
\item{obj}{an object of class \code{\link{pdmpBorder}}}

\item{value}{the value that shall be set}
}
\description{
erster satz...
}
