% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmpBorder_class.R
\docType{class}
\name{pdmpBorder-class}
\alias{pdmpBorder-class}
\alias{pdmp}
\alias{pdmpborder}
\alias{PDMP}
\alias{pdmpBorder}
\title{Class pdmpBorder}
\usage{
pdmpBorder(obj = NULL, descr = character(0), dynfunc, jumpfunc,
  ratefunc, borderfunc = function(t, z, parms) z, borroot = function(t,
  z, parms) NULL, terroot = function(t, z, parms) NULL, times = c(from
  = 0, to = 10, by = 1), init = c(0, 0), parms = c(0),
  discStates = list(0), out = NULL, solver = "lsodar",
  initfunc = NULL)
}
\arguments{
\item{borderfunc}{a \code{function(t, x, parms)} that returns the new state after reaching a borroot.}

\item{borroot}{a \code{function(t, x, parms)} that returns a zero for each continous 
variable reaching a border where the process forces a jump (maximum).}

\item{terroot}{a \code{function(t, x, parms)} that returns a zero for each contionous 
variable reaching a border where the process forces a termination (minimum).}
}
\description{
An S4 class to represent a piecewise deterministic markov 
process with Borders (PDMP).\cr
This class is based on the \pkg{pdmpsim} which itselve is based on \pkg{simecol} Package and provides a possibility 
to simulate piecewise deterministic markov processes with borders. These processes
are described in [Zei09] and [Ben+15]. It's closed to the original definition given in [Dav84], but one Restriction left: 
The number of continous variables should be independent of the 
state (of the discrete variable) of the process.
The class is based on the \code{\link[pdmpsim]{pdmpModel}} class of package
\pkg{pdmpsim} but introduces additional slots
}
\section{Slots}{

\describe{
\item{\code{borderfunc}}{a \code{function(t, x, parms)} that returns the new state after reaching a borroot.}

\item{\code{borroot}}{a \code{function(t, x, parms)} that returns a zero for each continous 
variable reaching a border where the process forces a jump (maximum).}

\item{\code{terroot}}{a \code{function(t, x, parms)} that returns a zero for each contionous 
variable reaching a border where the process forces a termination (minimum).}
}}

\references{
\tabular{ll}{
\eqn{\,}{ }[Dav84] \tab Davis, M. H. (1984). Piecewise-deterministic Markov 
processes: A general class of \cr
 \tab non-diffusion stochastic models. \emph{Journal of the Royal 
 Statistical Society. Series B} \cr
 \tab \emph{(Methodological)}, 353-388. \cr
[Zei09] \tab S. Zeiser. \emph{Classical and Hybrid Modeling of Gene 
Regulatory Networks}. 2009. \cr
[Ben+15]\eqn{\,\,\,\,}{    } \tab Benaïm, M., Le Borgne, S., Malrieu, F., 
& Zitt, P. A. (2015). Qualitative properties \cr
\tab of certain piecewise deterministic Markov processes. 
In \emph{Annales de l'Institut Henri} \cr
\tab \emph{Poincaré, Probabilités et Statistiques} (Vol. 51, No. 3, 
pp. 1040-1075). Institut \cr
\tab Henri Poincaré. \cr
 }
}
\seealso{
Class \code{pdmpModel} provides a method \code{\link{sim}} for simulation, 
\code{\link{pdmpBorder-accessors}{accessor functions}}
}
