% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjp_methods.R
\docType{methods}
\name{plot,mjpModel,missing-method}
\alias{plot,mjpModel,missing-method}
\title{Plot a PDMP}
\usage{
\S4method{plot}{mjpModel,missing}(x, y, ...)
}
\arguments{
\item{x}{an object of class mjpModel with a simulation 
stored in slot \code{out}}

\item{y}{ignored}

\item{...}{optional plotting parameters}
}
\description{
This is method plots single simulations of
markov jump processes defined as
\code{\link{mjpModel}}. There are also other plot
methods available that use \pkg{ggplot2}.
}
\examples{
data("toggleSwitch")
sim <- sim(toggleSwitch, seed = 1)
plot(sim, col = "red", lwd = 2)

# Alternative: plotSeeds
msim <- multSim(toggleSwitch, seeds = 1)
plot <- plotSeeds(msim)
plot + ggplot2::facet_grid(variable ~ seed)
}
\seealso{
\code{\link{plotSeeds}} for another plot function
to plot single simulations
}
