% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmp_vd_methods.R
\docType{methods}
\name{plot,pdmp_vd_Model,missing-method}
\alias{plot,pdmp_vd_Model,missing-method}
\title{Plot a PDMP}
\usage{
\S4method{plot}{pdmp_vd_Model,missing}(x, y, ...)
}
\arguments{
\item{x}{an object of class pdmpModel with a simulation 
stored in slot \code{out}}

\item{y}{ignored}

\item{...}{optional plotting parameters}
}
\description{
This is method plots single simulations of
piecewise deterministic markov processes defined as
\code{\link{pdmpModel}}. There are also other plot
methods available that use \pkg{ggplot2}.
}
\examples{
data("IASP")
sim <- sim(IASP, seed = 1)
plot(sim, col = "red", lwd = 2)

}
\seealso{
\code{\link{plotSeeds}} for another plot function, 
not yet available for variable dimension
}
