% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmpBorder_methods.R
\docType{methods}
\name{plot,pdmpBorder,missing-method}
\alias{plot,pdmpBorder,missing-method}
\title{Plot a PDMP}
\usage{
\S4method{plot}{pdmpBorder,missing}(x, y, ...)
}
\arguments{
\item{x}{an object of class pdmpBorder with a simulation 
stored in slot \code{out}}

\item{y}{ignored}

\item{...}{optional plotting parameters}
}
\description{
This is method plots single simulations of
piecewise deterministic markov processes defined as
\code{\link{pdmpBorder}}. There are also other plot
methods available that use \pkg{ggplot2}.
}
\seealso{
\code{\link{plotSeeds}} for another plot function
to plot single simulations
}
