% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSimData_outputMethods.R,
%   R/multSim_outputMethods.R
\name{plot}
\alias{plot}
\alias{plot.multSimData}
\alias{plot.multSim}
\title{Plot a 2d heatmap for multiple simulations of PDMPs}
\usage{
\method{plot}{multSimData}(x, title = NULL, subtitle = NULL,
  discPlot = "smooth", contPlot = "bin2d", ...)

\method{plot}{multSim}(x, title, subtitle, ...)
}
\arguments{
\item{x}{an object of class \code{\link{multSim}} or
\code{\link{multSimData}}}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}

\item{discPlot}{character vector determining the type of the plot
for the discrete variables. Possible values are 'line' standing for 
geom_line and 'smooth' standing for geom_smooth. Defaults to 'smooth'.}

\item{contPlot}{character vector determining the type of the plot
for the continous variables. Possible values are 'bin2d' standing for 
geom_bin2d and 'density_2d' standing for geom_density_2d.
Defaults to 'bin2d'.}

\item{...}{optional parameters to geom_bin2d resp. geom_density_2d 
and geom_line resp. geom_smooth}
}
\description{
This method plots multiple simulations of PDMPs.
Values of the continous variables are plotted in a 2d heatmap
whereas values of discrete variables are plotted as functions.
}
\examples{
data("simplePdmp")
ms <- multSim(simplePdmp, 1:20)
}
\seealso{
\code{\link{plotSeeds}}, \code{\link{plotTimes}}, 
\code{\link{plotStats}}, \code{\link{hist}} and \code{\link{density}} 
for other plot functions
}
