% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSimData_outputMethods.R,
%   R/multSim_outputMethods.R
\name{plotSeeds}
\alias{plotSeeds}
\alias{plotSeeds.multSimData}
\alias{plotSeeds.multSim}
\title{Plot method for simulations}
\usage{
plotSeeds(x, seeds, ...)

\method{plotSeeds}{multSimData}(x, seeds = NULL, ...)

\method{plotSeeds}{multSim}(x, seeds, ...)
}
\arguments{
\item{x}{object of class \code{\link{multSim}} or \code{\link{multSimData}}.}

\item{seeds}{vector with seed numbers to plot. This is optional if \code{x}
is a \code{\link{multSimData}} object.}

\item{...}{additional arguments, currently not used.}
}
\description{
This method plots the simulation for every seed in the input. 
Continous variables are plotted as line plots, 
discrete variables in a stacked barplot beneath.
The plot is created with \pkg{ggplot2} and can be modified afterwards.
}
\note{
A maximal number of 4 seeds can be plotted. 
The method requires the package \pkg{tidyr}.
}
\examples{
data("toggleSwitch")
msim <- multSim(toggleSwitch, seeds = 1:2)
plotSeeds(msim)
plotSeeds(msim, seeds = 1) + ggplot2::facet_grid(variable ~ seed)
}
