% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSimData_outputMethods.R,
%   R/multSim_outputMethods.R
\name{plotStats}
\alias{plotStats}
\alias{plotStats.multSimData}
\alias{plotStats.multSim}
\title{Plot statistics of multiple simulations}
\usage{
plotStats(x, vars, funs, ...)

\method{plotStats}{multSimData}(x, vars, funs, ...)

\method{plotStats}{multSim}(x, vars, funs, ...)
}
\arguments{
\item{x}{object of class \code{\link{multSim}} or \code{\link{multSimData}}.}

\item{vars}{character vector giving the names of variables to plot the 
statistics for.}

\item{funs}{a list of function calls generated by \code{\link[dplyr]{funs}}, 
or a character vector of function names, or simply a function.}

\item{...}{Additional arguments for the function calls in funs.}
}
\description{
This method plots descriptive statistics as i. e. mean, min, max, sd
over all simulations depending on the time. The desired statistic values
can be specified individually with the parameter \code{funs}.
}
\note{
The method requires the packages \pkg{grid} and \pkg{gtable}.
}
\examples{
data("toggleSwitch")
ms <- multSim(toggleSwitch, seeds = 1:10)
plotStats(ms, vars = c("fA", "fB"), funs = "mean")
plotStats(ms, vars = "fB", funs = c("min", "max", "mean"))
}
\seealso{
\link{summarise_at} to get the calculated values of the statistics
}
