% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mjp_methods.R
\docType{methods}
\name{print,mjpModel-method}
\alias{print,mjpModel-method}
\title{Methods for funktion print in package \pkg{pdmpsim}}
\usage{
\S4method{print}{mjpModel}(x, all = FALSE, ...)
}
\arguments{
\item{all}{specifies whether all slots are printed. If FALSE (the default),
only the slots that characterize the model will be printed.}

\item{...}{optional parameters passed to print.}
}
\description{
Methods for funktion print in package \pkg{pdmpsim}
}
