% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmp_vd_methods.R
\docType{methods}
\name{print,pdmp_vd_Model-method}
\alias{print,pdmp_vd_Model-method}
\title{Methods for funktion print in package \pkg{pdmpsim}}
\usage{
\S4method{print}{pdmp_vd_Model}(x, all = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{pdmp_vd_Model} or one of its subclasses.}

\item{all}{speciefies whether all slots are printed. If FALSE (the default),
only the slots that characterize the model will be printed.}

\item{...}{optional parameters passed to print.}
}
\description{
Methods for funktion print in package \pkg{pdmpsim}
}
