% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSim_calculations.R
\name{removeSeeds}
\alias{removeSeeds}
\alias{removeseeds}
\title{Remove not simulated seeds}
\usage{
removeSeeds(ms)
}
\arguments{
\item{ms}{object of class \code{\link{multSim}}}
}
\value{
object of class \code{multSim} that does not contain NAs in slot 
\code{outputList}
}
\description{
If a simulation via \code{\link{multSim}} has been interrupted, slot 
\code{outputList} of the returned \code{multSim} object contains NAs. 
Method \code{removeSeeds} removes these NA entries and changes all other 
slots accordingly. This method is used internally for other calculation 
methods as e.g. \code{\link{mean}}.
}
\examples{
data(simplePdmp)
setTimeLimit(cpu = 2)
ms <- multSim(simplePdmp, seeds = 1:100)
removeSeeds(ms)
}
