% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SIRstoch}
\alias{SIRstoch}
\title{Stochastic SIR dynamics}
\format{An object of class \code{\link{mjpModel}}.}
\usage{
SIRstoch
}
\description{
This is a more complicated example for a markov jump process
defined as \code{\link{mjpModel}}. It is included to the package for 
demonstration purposes.
}
\section{Slots}{

\describe{
\item{\code{parms}}{infection (\code{beta}) and recovery  rate (\code{gamma}).}

\item{\code{init}}{There are three integer  variables \code{S,I,R} with initial value (10,1,0).}

\item{\code{jumpfunc}}{There are two jumptypes. infections  jump from \code{(S,I,R)} to 
\code{(S-1,I+1,R)}, recoveries from \code{(S,I,R)} to \code{(S,I-1,R+1)}.}

\item{\code{ratefunc}}{A vector of length two determining the probability of a jump 
being a recovery or infection.}

\item{\code{times}}{The simulations will start at time \code{t = 0} and end at 
\code{t = 10} with step length 0.01.}
}}

\examples{
## the code used to generate this model:

SIRstoch <- mjpModel(
    descr = "stochastic SIR epidemic",
    init = c(S=10,I=1,R=0),
    times = c(from = 0, to = 10, by = 0.01),
    ratefunc = function(t, x, parms) c(parms[1]*x[1]*x[2],parms[2]*x[2]),
    jumpfunc = function(t, x, parms, jtype){
         x+switch(jtype,c(-1,1,0),c(0,-1,1))
    },
    parms=c(beta=1,gamma=1))

## load it and plot a simulation:
data("SIRstoch")
plot(sim(SIRstoch))
}
\seealso{
\code{\link{KendallBD}} for a  simpler example of a 
\code{mjpModel} and \code{\link{mjpModel-class}}
for the formal description of the S4 class.
}
\keyword{datasets}
