% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msCsv_calculations.R, R/multSim_calculations.R
\name{mean.multSimCsv}
\alias{mean.multSimCsv}
\alias{mean}
\alias{mean.multSim}
\title{Methods for function \code{mean} in Package \pkg{pdmpsim}}
\usage{
\method{mean}{multSimCsv}(x, ...)

\method{mean}{multSim}(x, ...)
}
\arguments{
\item{x}{object of class \code{\link{multSim}} or \code{\link{multSimCsv}}}

\item{...}{additional arguments passed to \code{\link[base]{mean}}}
}
\value{
data.frame with calculated mean values
}
\description{
This method calculates the mean for every time value over all simulations.
}
\note{
Methods \code{mean.multSim} and \code{mean.multSimCsv} can lead to 
slightly different results. The reason for this is that the simulation results 
created by \code{\link{multSimCsv}} are stored with a fewer number of digits.
}
\examples{
data(simplePdmp)
sim <- multSim(simplePdmp, seeds = 1:10)
mean(sim)

# handling NA values:
sim$outputList[[1]][1,2] <- NA
sim$outputList <- c(sim$outputList, NA) # will be removed before calculation
head(mean(sim))
head(mean(sim, na.rm = TRUE))

# Methods mean.multSim and mean.multSimCsv
# can lead to slightly different results:
sim2 <- multSimCsv(simplePdmp, seeds = 1:10)
all.equal(mean(sim), mean(sim2))
}
