% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmp_accessors.R
\docType{methods}
\name{pdmp-accessors}
\alias{pdmp-accessors}
\alias{dynfunc}
\alias{dynfunc<-}
\alias{ratefunc}
\alias{ratefunc<-}
\alias{jumpfunc}
\alias{jumpfunc<-}
\alias{descr}
\alias{descr<-}
\alias{parms<-}
\alias{discStates}
\alias{discStates<-}
\alias{dynfunc,pdmpModel-method}
\alias{ratefunc,pdmpModel-method}
\alias{jumpfunc,pdmpModel-method}
\alias{descr,pdmpModel-method}
\alias{discStates,pdmpModel-method}
\alias{dynfunc<-,pdmpModel-method}
\alias{ratefunc<-,pdmpModel-method}
\alias{jumpfunc<-,pdmpModel-method}
\alias{parms<-,pdmpModel-method}
\alias{discStates<-,pdmpModel-method}
\alias{descr<-,pdmpModel-method}
\alias{main<-,pdmpModel-method}
\alias{equations<-,pdmpModel-method}
\alias{observer<-,pdmpModel-method}
\title{Accessor functions for Class pdmpModel}
\usage{
dynfunc(obj)

dynfunc(obj) <- value

ratefunc(obj)

ratefunc(obj) <- value

jumpfunc(obj)

jumpfunc(obj) <- value

descr(obj)

descr(obj) <- value

parms(obj) <- value

discStates(obj)

discStates(obj) <- value

\S4method{dynfunc}{pdmpModel}(obj)

\S4method{ratefunc}{pdmpModel}(obj)

\S4method{jumpfunc}{pdmpModel}(obj)

\S4method{descr}{pdmpModel}(obj)

\S4method{discStates}{pdmpModel}(obj)

\S4method{dynfunc}{pdmpModel}(obj) <- value

\S4method{ratefunc}{pdmpModel}(obj) <- value

\S4method{jumpfunc}{pdmpModel}(obj) <- value

\S4method{parms}{pdmpModel}(obj) <- value

\S4method{discStates}{pdmpModel}(obj) <- value

\S4method{descr}{pdmpModel}(obj) <- value

\S4method{main}{pdmpModel}(obj) <- value

\S4method{equations}{pdmpModel}(obj) <- value

\S4method{observer}{pdmpModel}(obj) <- value
}
\arguments{
\item{obj}{an object of class \code{\link{pdmpModel}}}

\item{value}{the value that shall be set}
}
\description{
Get or set the slot values for an object of class \code{\link{pdmpModel}}. 
In most cases, setting a new value also affects the slot \code{out} 
(where the simulation result can be stored). This applies to slots 
\code{times}, \code{parms}, \code{init}, \code{dynfunc}, \code{ratefunc},
\code{jumpfunc} and \code{initfunc}, which set out to NULL.
Setting a new value for \code{descr} or \code{discStates} will not affect 
the slot \code{out}.
}
\details{
It is also possible to modify the slots directly,
e.g. the parameters of a model \code{pdmp} with \code{pdmp@parms}, 
but this is normally not recommended as there is no guarantee that this 
will work in a compatible way in future versions.
}
\seealso{
Class definition \code{\link{pdmpModel}}
}
