% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmp_methods.R
\docType{methods}
\name{plot,pdmpModel,missing-method}
\alias{plot,pdmpModel,missing-method}
\title{Plot a PDMP}
\usage{
\S4method{plot}{pdmpModel,missing}(x, y, ...)
}
\arguments{
\item{x}{an object of class pdmpModel with a simulation 
stored in slot \code{out}}

\item{y}{ignored}

\item{...}{optional plotting parameters}
}
\description{
These methods plot single simulations of
piecewise deterministic Markov processes defined as
\code{\link{pdmpModel}} or \code{\link{pdmp_vd_Model}}. Note that there are further plot
methods implemented that use \pkg{ggplot2}.
}
\examples{
data("toggleSwitch")
sim <- sim(toggleSwitch, seed = 1)
plot(sim, col = "red", lwd = 2)

# Alternative: plotSeeds
msim <- multSim(toggleSwitch, seeds = 1)
plot <- plotSeeds(msim)
plot + ggplot2::facet_grid(variable ~ seed)
# variable dimensions: summary functions apply
data("IASP")
sim <- sim(IASP, seed = 1)
plot(sim, col = "red", lwd = 2)
}
\seealso{
\code{\link{plotSeeds}} for another plot function
to plot single simulations
}
