% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmp_methods.R
\docType{methods}
\name{plot,pdmpModel,missing-method}
\alias{plot,pdmpModel,missing-method}
\title{Plot a PDMP}
\usage{
\S4method{plot}{pdmpModel,missing}(x, y, ggplot = FALSE, ...)
}
\arguments{
\item{x}{an object of class pdmpModel with a simulation 
stored in slot \code{out}}

\item{y}{ignored}

\item{ggplot}{boolean variable. If TRUE, a different plot 
will be created that uses package \pkg{ggplot2}.}

\item{...}{optional parameters to plot (if ggplot = FALSE)
or getMultSimData (if ggplot = TRUE)}
}
\description{
This is method plots single simulations of
piecewise deterministic markov processes defined as
\code{\link{pdmpModel}}. There are also other plot
methods available that use \pkg{ggplot2}.
}
\examples{
data("toggleSwitch")
sim <- sim(toggleSwitch, seed = 1)
plot(sim, col = "red", lwd = 2)
plot(sim, ggplot = TRUE)
plot(sim, ggplot = TRUE, seed = 1) + ggplot2::facet_grid(variable ~ seed)
}
\seealso{
\code{\link{plotSeeds}} for another plot function
to plot single simulations
}
