% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSimData_outputMethods.R,
%   R/multSim_outputMethods.R
\name{plotStats}
\alias{plotStats}
\alias{plotStats.multSimData}
\alias{plotStats.multSim}
\title{Plot statistics of multiple simulations}
\usage{
plotStats(x, vars, funs, plottype, ...)

\method{plotStats}{multSimData}(x, vars, funs, plottype = "line", ...)

\method{plotStats}{multSim}(x, vars, funs, ...)
}
\arguments{
\item{x}{object of class \code{\link{multSim}} or \code{\link{multSimData}}.}

\item{vars}{character vector giving the names of variables to plot the 
statistics for.}

\item{funs}{a list of function calls generated by \code{\link[dplyr]{funs}}, 
or a character vector of function names, or simply a function.}

\item{plottype}{a string indicating which plotmethod should be used.
Either "line" for  \code{\link[ggplot2]{geom_line}} or "smooth" for
\code{\link[ggplot2]{geom_smooth}} are possible. Other values are ignored.}

\item{...}{Additional arguments for the function calls in funs.}
}
\description{
This method plots descriptive statistics as i. e. mean, min, max, sd
over all simulations depending on the time. The desired statistic values
can be specified individually with the parameter \code{funs}. The plot is 
created with \pkg{ggplot2} and can be modified afterwards.
}
\examples{
data("toggleSwitch")
ms <- multSim(toggleSwitch, seeds = 1:10)
plotStats(ms, vars = c("fA", "fB"), funs = "mean")
plotStats(ms, vars = "fB", funs = c("min", "max", "mean"))
}
\seealso{
\link{summarise_at} to get the calculated values of the statistics
}
