% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multSimData_outputMethods.R,
%   R/multSim_outputMethods.R
\name{plotTimes}
\alias{plotTimes}
\alias{plotTimes.multSimData}
\alias{plotTimes.multSim}
\title{Plot at different times}
\usage{
plotTimes(x, vars, times, nolo = 0, plottype = "boxplot", ...)

\method{plotTimes}{multSimData}(x, vars, times, nolo = 0,
  plottype = "boxplot", ...)

\method{plotTimes}{multSim}(x, vars, times, nolo = 0, plottype = "boxplot",
  ...)
}
\arguments{
\item{x}{object of class \code{\link{multSim}} or \code{\link{multSimData}}.}

\item{vars}{character vector of variable names that shall be plotted}

\item{times}{numeric vector with time values to plot. If no vector is given,
ten values between the minimum and maximum provided time values will be 
plotted.}

\item{nolo}{an integer giving the number of outliers that shall be
labelled with the seed number that was used for simulation. Defaults to 0.}

\item{plottype}{character vector determining the type of the plot.
Possible values are 'boxplot', 'violin' or 'dotplot'. Defaults to 'boxplot'.}

\item{...}{additional parameters for the plotting function (either 
geom_boxplot, geom_violin or geom_dotplot, depending on parameter plottype).}
}
\description{
Plot a boxplot or violin plot or dotplot for every time value 
in the data.frame (a maximum of 12 different time values is allowed).
Outliers can be plotted with a number on their side which represents the 
seed number that was #' used to simulate the simulation that causes the 
outlier. The number of outliers that shall be plotted this way is determined
by parameter \emph{nolo}(= number of labelled outliers).
A red diamond represents the median of the simulated values.
The plot is created with \pkg{ggplot2} and can be modified afterwards.
}
\examples{
data("simplePdmp")
times(simplePdmp) <- c(from = 0, to = 5, by = 1)
md <- getMultSimData(multSim(simplePdmp, 1:8), times = 1:5)
plotTimes(md, plottype = 'violin')
plotTimes(md, nolo = 3)
}
